refactor(mm-next): rename filename to follow convention
