+++ packages/mirror-media-next/components/shared/layout-full.js
